+++ Changes.md
@@ -15,6 +15,7 @@ Note: development is tracked on the [`develop` branch](https://github.com/chmp/i
     - Use markdown for module documentation
     - Updated documentation (updated doc strings for run, ipytest magic,
       expanded example notebook)
+    - Fix bug for `--deselect {MODULE}::test`
 - `0.12.0`:
     - Re-add the `raise_on_error` config option
     - Return the `exit_code` from `ipyest.run()`